+++ src/test/java/com/epam/reportportal/cucumber/integration/feature/EmptySteps.java
@@ -2,8 +2,9 @@ package com.epam.reportportal.cucumber.integration.feature;
 
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
-import org.slf4j.LoggerFactory;
+import io.cucumber.java.en.When;
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class EmptySteps {
 	private static final Logger LOGGER = LoggerFactory.getLogger(EmptySteps.class);
@@ -16,5 +17,10 @@ public class EmptySteps {
 	@Then("I have another empty step")
 	public void i_have_another_empty_step() {
 		LOGGER.info("Inside 'I have another empty step'");
+	}
+
+	@When("I have one more empty step")
+	public void i_have_one_more_empty_step() {
+		LOGGER.info("I have one more empty step'");
 	}
 }