+++ src/test/java/com/epam/reportportal/cucumber/TestCaseIdTest.java
@@ -2,6 +2,7 @@ package com.epam.reportportal.cucumber;
 
 import com.epam.reportportal.cucumber.integration.TestScenarioReporter;
 import com.epam.reportportal.cucumber.integration.TestStepReporter;
+import com.epam.reportportal.cucumber.integration.feature.TestCaseIdOnMethodSteps;
 import com.epam.reportportal.cucumber.integration.util.TestUtils;
 import com.epam.reportportal.listeners.ListenerParameters;
 import com.epam.reportportal.service.ReportPortal;
@@ -41,6 +42,13 @@ public class TestCaseIdTest {
 			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
 			"com.epam.reportportal.cucumber.integration.TestStepReporter" })
 	public static class RunBellyTestStepReporter extends AbstractTestNGCucumberTests {
+
+	}
+
+	@CucumberOptions(features = "src/test/resources/features/TestCaseIdOnAMethod.feature", glue = {
+			"com.epam.reportportal.cucumber.integration.feature" }, plugin = { "pretty",
+			"com.epam.reportportal.cucumber.integration.TestStepReporter" })
+	public static class StepDefStepReporter extends AbstractTestNGCucumberTests {
 
 	}
 
@@ -101,6 +109,14 @@ public class TestCaseIdTest {
 		);
 
 	}
+
+	@Test
+	public void verify_test_case_id_bypassed_through_annotation_on_a_stepdef() {
+		TestUtils.runTests(StepDefStepReporter.class);
+		ArgumentCaptor<StartTestItemRQ> captor = ArgumentCaptor.forClass(StartTestItemRQ.class);
+		verify(client, times(1)).startTestItem(same(testId), captor.capture());
+
+		StartTestItemRQ step = captor.getValue();
+		assertThat(step.getTestCaseId(), equalTo(TestCaseIdOnMethodSteps.TEST_CASE_ID));
+	}
 }
-
-