+++ src/main/java/com/epam/reportportal/cucumber/ScenarioReporter.java
@@ -15,10 +15,12 @@
  */
 package com.epam.reportportal.cucumber;
 
+import com.epam.reportportal.service.Launch;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import io.cucumber.messages.Messages;
 import io.cucumber.plugin.event.*;
 import io.reactivex.Maybe;
+import org.apache.commons.lang3.tuple.Pair;
 import rp.com.google.common.base.Supplier;
 import rp.com.google.common.base.Suppliers;
 
@@ -47,14 +49,14 @@ import java.util.Calendar;
  */
 public class ScenarioReporter extends AbstractReporter {
     private static final String SEPARATOR = " --- ";
-    private static final String RP_TEST_TYPE = "TEST";
+    private static final String RP_STORY_TYPE = "SUITE";
+    private static final String RP_TEST_TYPE = "STORY";
     private static final String RP_STEP_TYPE = "STEP";
     private static final String HOOK_COLON = "Hook:";
     private static final String COLON_ = ": ";
     private static final String STEP_ = "STEP ";
     private static final String EMPTY_SUFFIX = "";
     private static final String INFO = "INFO";
-    private static final String RP_STORY_TYPE = "STORY";
     private static final String DUMMY_ROOT_SUITE_NAME = "Root User Story";
 
     protected Supplier<Maybe<String>> rootSuiteId;
@@ -67,38 +69,49 @@ public class ScenarioReporter extends AbstractReporter {
 
     @Override
     protected void beforeStep(TestStep testStep) {
-        RunningContext.ScenarioContext currentScenarioContext = getCurrentScenarioContext();
-        Messages.GherkinDocument.Feature.Step step = currentScenarioContext.getStep(testStep);
-        int lineInFeatureFile = step.getLocation().getLine();
-        String decoratedStepName = lineInFeatureFile + decorateMessage(Utils.buildNodeName(currentScenarioContext.getStepPrefix(),
-                step.getKeyword(),
-                Utils.getStepName(testStep),
-                EMPTY_SUFFIX
-        ));
-        String multilineArg = Utils.buildMultilineArgument(testStep);
-        Utils.sendLog(decoratedStepName + multilineArg, INFO);
+        RunningContext.ScenarioContext context = getCurrentScenarioContext();
+        Messages.GherkinDocument.Feature.Step step = context.getStep(testStep);
+        StartTestItemRQ rq = Utils.buildStartStepRequest(context.getStepPrefix(), testStep, step, false);
+        context.setCurrentStepId(launch.get().startTestItem(context.getId(), rq));
     }
 
     @Override
     protected void afterStep(Result result) {
-        if (result.getStatus() != Status.PASSED) {
-            reportResult(result, decorateMessage(STEP_ + result.getStatus().name().toUpperCase()));
-        }
+        reportResult(result, null);
+        RunningContext.ScenarioContext context = getCurrentScenarioContext();
+        Launch myLaunch = launch.get();
+        Utils.finishTestItem(myLaunch, context.getCurrentStepId(), result.getStatus());
+        context.setCurrentStepId(null);
+        myLaunch.getStepReporter().finishPreviousStep();
     }
 
     @Override
     protected void beforeHooks(HookType hookType) {
-        // noop
+        StartTestItemRQ rq = new StartTestItemRQ();
+        rq.setHasStats(false);
+        Pair<String, String> typeName = Utils.getHookTypeAndName(hookType);
+        rq.setType(typeName.getKey());
+        rq.setName(typeName.getValue());
+        rq.setStartTime(Calendar.getInstance().getTime());
+
+        RunningContext.ScenarioContext context = getCurrentScenarioContext();
+        context.setHookStepId(launch.get().startTestItem(getCurrentScenarioContext().getId(), rq));
+        context.setHookStatus(Status.PASSED);
     }
 
     @Override
     protected void afterHooks(Boolean isBefore) {
-        // noop
+        RunningContext.ScenarioContext context = getCurrentScenarioContext();
+        Launch myLaunch = launch.get();
+        Utils.finishTestItem(myLaunch, context.getHookStepId(), context.getHookStatus());
+        context.setHookStepId(null);
+        myLaunch.getStepReporter().finishPreviousStep();
     }
 
     @Override
     protected void hookFinished(HookTestStep step, Result result, Boolean isBefore) {
-        reportResult(result, step.getHookType().name() + HOOK_COLON + result.getStatus() + COLON_ + step.getCodeLocation());
+        reportResult(result, (isBefore ? "Before" : "After") + " hook: " + step.getCodeLocation());
+        getCurrentScenarioContext().setHookStatus(result.getStatus());
     }
 
     @Override